Silver can't handle annotated inner classes, so move these out to the top level
